Crate recess

An interface for interacting with the Rust playground

examples

// recess interfaces
extern crate recess;
// tokio async io
extern crate tokio_core;
// futures combinators
extern crate futures;

use recess::{Client, CompileRequest};
use futures::Future;
use tokio_core::reactor::Core;

fn main() {
  let mut core = Core::new().unwrap();
  let client = Client::new(
     &core.handle(),
  );

  let work = client.compile(CompileRequest::builder(
             r#"fn main() { println!("{}", 1); }"#
           )
          .build().unwrap())
       .and_then(|result| {
           println!("{}", result.stdout);
           println!("{}", result.stderr);
           Ok(())
       });

  println!("{:#?}", core.run(work))
}

Cargo features

This crate has one Cargo feature, tls, which adds HTTPS support via the Client::new constructor. This feature is enabled by default.

Reexports

pub use compile::Request as CompileRequest;

pub use execute::Request as ExecuteRequest;

pub use format::Request as FormatRequest;

pub use lint::Request as LintRequest;

Modules

compile
Compilation interfaces

execute
Execution interfaces

format
Format interfaces

lint
Linting interfaces

Structs

Client
Rust playground client

Error
The Error type.

Enums

AsmFlavor
Assembly flavor.

Backtrace
Rustc backtrace options

Channel
Release train options.

CompileOutput
Compiler output formats

CrateType
Type of crate

ErrorKind
The kind of an error.

Mode
Rustc compilation mode.

OptLevel
Optimization levels for rustc

Traits

ResultExt
Additional methods for Result, for easy interaction with this crate.

Type Definitions

Future
A type alias for futures that may return recess::Error's

Result
Convenient wrapper around std::Result.